+++ src/main/kotlin/backend/itracker/crawl/service/CrawlService.kt
@@ -1,7 +1,7 @@
 package backend.itracker.crawl.service
 
-import backend.itracker.crawl.domain.Macbook
-import backend.itracker.crawl.service.common.Crawler
+import backend.itracker.crawl.macbook.domain.Macbook
+import backend.itracker.crawl.service.util.Crawler
 import org.springframework.stereotype.Component
