+++ src/main/kotlin/backend/itracker/crawl/service/util/helper/DriverConnector.kt
@@ -1,4 +1,4 @@
-package backend.itracker.crawl.service.common
+package backend.itracker.crawl.service.util.helper
 
 import org.openqa.selenium.WebDriver
 import org.openqa.selenium.chrome.ChromeDriver
@@ -12,7 +12,7 @@ private const val USER_AGENT =
     "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/124.0.0.0 Safari/537.36"
 private const val DEFAULT_WAIT_TIME = 5L
 
-class DriverHelper {
+class DriverConnector {
 
     fun getDriver(): RemoteWebDriver {
         val chromeOptions = ChromeOptions()
