+++ src/test/resources/application.yml
@@ -1,0 +1,10 @@
+spring:
+  application:
+    name: itracker
+  jpa:
+    hibernate:
+      ddl-auto: create
+    properties:
+      hibernate:
+        default_batch_fetch_size: 100
+        format_sql: true
